Product list landing (#15)
* feat: add manage category page

* feat: add create product page

* feat: add feature delete product

* fix: admin sidebar component

* feat: add manage store and manage admin page

* fix: fix get all data of users on user.router

* fix: add paranoid option on user model

* fix: change background color on all management page

* feat: add feature edit profile admin and change password admin

* fix: restrict access store admin on manage category and product pages

* fix: add coloumn unit to table product

* fix: fix bug manage product

* feat: add search feature on landing page

* fix: change parameter name on resTemplate.js

+++ apps/api/src/controllers/category.controller.js
@@ -1,10 +1,26 @@
+import { Op } from "sequelize";
 import category from "../models/categories.model";
-import { unlink } from "fs";
+import { unlink, existsSync } from "fs";
 
 const dir = './src/assets/category/';
 
 export const createCategory = async (data, image) => {
-  const value = image ? {name: data.name, image: image.filename} : {name: data.name};
+  const checkCategory = await category.findOne({
+    where: {
+      name: data.name,
+    }
+  });
+  
+  if (checkCategory) {
+    throw {
+      rc: 409,
+      success: false,
+      message: 'Category already exists',
+      result: null,
+    }
+  }
+
+  const value = image ? { name: data.name, image: image.filename } : { name: data.name };
   return await category.create(value);
 };
 
@@ -16,37 +32,55 @@ export const getCategoryDataById = async (id) => {
   return await category.findByPk(id);
 };
 
-export const updateCategory = async (id, data, image) => { 
-  if(image){
+export const updateCategory = async (id, data, image) => {
+  console.log('DATA >>> ', data);
+  console.log('DATA >>> ', Object.hasOwn(data, 'name'));
+  if(Object.hasOwn(data, 'name')){
+    const checkCategory = await category.findOne({ where: {name: data.name} });
+    if (checkCategory) {
+      throw {
+        rc: 409,
+        success: false,
+        message: 'Category already exists',
+        result: null,
+      }
+    }
+  }
+
+  if (image) {
     const prevData = await category.findByPk(id);
 
     await category.update(
-      { 
-        name: data.name, 
-        image: image.filename 
-      }, 
-      { where: { id } 
-    });
+      Object.hasOwn(data, 'name') ? {name: data.name,image: image.filename} : {image: image.filename}
+      ,{
+        where: { id }
+      });
 
-    unlink(dir + prevData.image, (err) => {
-      if (err) throw Error(err);
-    });
+    if (existsSync(dir + prevData.image)) {
+      console.log('updating image');
+      unlink(dir + prevData.image, (err) => {
+        if (err) throw Error(err);
+      });
+    }
   }
 
   await category.update(
-    { 
+    {
       name: data.name
-    }, 
-    { where: { id } 
-  });
+    },
+    {
+      where: { id }
+    });
 };
 
-export const deleteCategory = async (id) => { 
+export const deleteCategory = async (id) => {
   const prevData = await category.findByPk(id);
 
   await category.destroy({ where: { id } });
 
-  unlink(dir + prevData.image, (err) => {
-    if (err) throw Error(err);
-  });
+  if (existsSync(dir + prevData.image)) {
+    unlink(dir + prevData.image, (err) => {
+      if (err) throw Error(err);
+    });
+  }
 };
